Fix bugs in ReadVtkStructuredPoints filter
