+++ src/Plugins/SimplnxCore/src/SimplnxCore/Filters/Algorithms/WriteVtkRectilinearGrid.cpp
@@ -5,218 +5,14 @@
 #include "simplnx/Utilities/FilterUtilities.hpp"
 #include "simplnx/Utilities/StringUtilities.hpp"
 
+#include "SimplnxCore/utils/VtkUtilities.hpp"
+
 #include <H5Tpublic.h>
 
 using namespace nx::core;
 
 namespace
 {
-static constexpr usize k_BufferDumpVal = 1000000;
-
-// -----------------------------------------------------------------------------
-template <typename T>
-std::string TypeForPrimitive(T value, const IFilter::MessageHandler& messageHandler)
-{
-  if constexpr(std::is_same_v<T, float32>)
-  {
-    return "float";
-  }
-  if constexpr(std::is_same_v<T, float64>)
-  {
-    return "double";
-  }
-
-  if constexpr(std::is_same_v<T, int8>)
-  {
-    return "char";
-  }
-  if constexpr(std::is_same_v<T, uint8>)
-  {
-    return "unsigned_char";
-  }
-  if constexpr(std::is_same_v<T, char>)
-  {
-    return "char";
-  }
-  if constexpr(std::is_same_v<T, signed char>)
-  {
-    return "char";
-  }
-  if constexpr(std::is_same_v<T, unsigned char>)
-  {
-    return "char";
-  }
-
-  if constexpr(std::is_same_v<T, int16>)
-  {
-    return "short";
-  }
-  if constexpr(std::is_same_v<T, short>)
-  {
-    return "short";
-  }
-  if constexpr(std::is_same_v<T, signed short>)
-  {
-    return "short";
-  }
-  if constexpr(std::is_same_v<T, uint16>)
-  {
-    return "unsigned_short";
-  }
-  if constexpr(std::is_same_v<T, unsigned short>)
-  {
-    return "unsigned_short";
-  }
-
-  if constexpr(std::is_same_v<T, int32>)
-  {
-    return "int";
-  }
-  if constexpr(std::is_same_v<T, uint32>)
-  {
-    return "unsigned_int";
-  }
-  if constexpr(std::is_same_v<T, int>)
-  {
-    return "int";
-  }
-  if constexpr(std::is_same_v<T, signed int>)
-  {
-    return "int";
-  }
-  if constexpr(std::is_same_v<T, unsigned int>)
-  {
-    return "unsigned_int";
-  }
-
-  if constexpr(std::is_same_v<T, long int>)
-  {
-    return "long";
-  }
-  if constexpr(std::is_same_v<T, signed long int>)
-  {
-    return "long";
-  }
-  if constexpr(std::is_same_v<T, unsigned long int>)
-  {
-    return "unsigned_long";
-  }
-
-  if constexpr(std::is_same_v<T, long long int>)
-  {
-    return "long";
-  }
-  if constexpr(std::is_same_v<T, signed long long int>)
-  {
-    return "long";
-  }
-  if constexpr(std::is_same_v<T, unsigned long long int>)
-  {
-    return "unsigned_long";
-  }
-  if constexpr(std::is_same_v<T, int64>)
-  {
-    return "long";
-  }
-  if constexpr(std::is_same_v<T, uint64>)
-  {
-    return "unsigned_long";
-  }
-
-  if constexpr(std::is_same_v<T, bool>)
-  {
-    return "char";
-  }
-
-  messageHandler(IFilter::Message::Type::Info, fmt::format("Error: TypeForPrimitive - Unknown Type: ", typeid(value).name()));
-  if(const char* name = typeid(value).name(); nullptr != name && name[0] == 'l')
-  {
-    messageHandler(
-        IFilter::Message::Type::Info,
-        fmt::format(
-            "You are using 'long int' as a type which is not 32/64 bit safe. It is suggested you use one of the H5SupportTypes defined in <Common/H5SupportTypes.h> such as int32_t or uint32_t.",
-            typeid(value).name()));
-  }
-  return "";
-}
-
-// -----------------------------------------------------------------------------
-struct WriteVtkDataArrayFunctor
-{
-  template <typename T>
-  void operator()(FILE* outputFile, bool binary, DataStructure& dataStructure, const DataPath& arrayPath, const IFilter::MessageHandler& messageHandler)
-  {
-    auto& dataArray = dataStructure.getDataRefAs<DataArray<T>>(arrayPath);
-
-    messageHandler(IFilter::Message::Type::Info, fmt::format("Writing Cell Data {}", dataArray.getName()));
-
-    const usize totalElements = dataArray.getSize();
-    const int numComps = static_cast<int>(dataArray.getNumberOfComponents());
-    std::string dName = dataArray.getName();
-    dName = StringUtilities::replace(dName, " ", "_");
-
-    const std::string vtkTypeString = TypeForPrimitive<T>(dataArray[0], messageHandler);
-    bool useIntCast = false;
-    if(vtkTypeString == "unsigned_char" || vtkTypeString == "char")
-    {
-      useIntCast = true;
-    }
-
-    fprintf(outputFile, "SCALARS %s %s %d\n", dName.c_str(), vtkTypeString.c_str(), numComps);
-    fprintf(outputFile, "LOOKUP_TABLE default\n");
-    if(binary)
-    {
-      if constexpr(endian::little == endian::native)
-      {
-        dataArray.byteSwapElements();
-      }
-      fwrite(dataArray.template getIDataStoreAs<DataStore<T>>()->data(), sizeof(T), totalElements, outputFile);
-      fprintf(outputFile, "\n");
-      if constexpr(endian::little == endian::native)
-      {
-        dataArray.byteSwapElements();
-      }
-    }
-    else
-    {
-      std::string buffer;
-      buffer.reserve(k_BufferDumpVal);
-      for(size_t i = 0; i < totalElements; i++)
-      {
-        if(i % 20 == 0 && i > 0)
-        {
-          buffer.append("\n");
-        }
-        if(useIntCast)
-        {
-          buffer.append(fmt::format(" {:d}", static_cast<int>(dataArray[i])));
-        }
-        else if constexpr(std::is_same_v<T, float32>)
-        {
-          buffer.append(fmt::format(" {:f}", dataArray[i]));
-        }
-        else if constexpr(std::is_same_v<T, float64>)
-        {
-          buffer.append(fmt::format(" {:f}", dataArray[i]));
-        }
-        else
-        {
-          buffer.append(fmt::format(" {}", dataArray[i]));
-        }
-        // If the buffer is within 32 bytes of the reserved size, then dump
-        // the contents to the file.
-        if(buffer.size() > (k_BufferDumpVal - 32))
-        {
-          fprintf(outputFile, "%s", buffer.c_str());
-          buffer.clear();
-          buffer.reserve(k_BufferDumpVal);
-        }
-      }
-      buffer.append("\n");
-      fprintf(outputFile, "%s", buffer.c_str());
-    }
-  }
-};
 } // namespace
 
 // -----------------------------------------------------------------------------
