+++ src/Plugins/SimplnxCore/src/SimplnxCore/Filters/ReadVtkStructuredPointsFilter.cpp
@@ -10,56 +10,15 @@
 #include "simplnx/Parameters/DataGroupCreationParameter.hpp"
 #include "simplnx/Parameters/DataObjectNameParameter.hpp"
 #include "simplnx/Parameters/FileSystemPathParameter.hpp"
-#include "simplnx/Utilities/VtkLegacyFileReader.hpp"
+#include "simplnx/Utilities/TimeUtilities.hpp"
 
-#include <chrono>
-#include <iostream>
-#include <iomanip> // For std::setw and std::setfill
 #include <filesystem>
 namespace fs = std::filesystem;
 
 using namespace nx::core;
 
-namespace {
-
-
-class StopWatch {
-public:
-  void start() {
-    start_time = std::chrono::steady_clock::now();
-  }
-
-  void stop() {
-    end_time = std::chrono::steady_clock::now();
-  }
-
-  void print(std::ostream& os) const {
-    auto elapsed = end_time - start_time;
-    auto hours = std::chrono::duration_cast<std::chrono::hours>(elapsed);
-    auto minutes = std::chrono::duration_cast<std::chrono::minutes>(elapsed % std::chrono::hours(1));
-    auto seconds = std::chrono::duration_cast<std::chrono::seconds>(elapsed % std::chrono::minutes(1));
-    auto milliseconds = std::chrono::duration_cast<std::chrono::milliseconds>(elapsed % std::chrono::seconds(1));
-
-    os << std::setw(2) << std::setfill('0') << hours.count() << ":"
-       << std::setw(2) << std::setfill('0') << minutes.count() << ":"
-       << std::setw(2) << std::setfill('0') << seconds.count() << "."
-       << std::setw(3) << std::setfill('0') << milliseconds.count();
-  }
-
-private:
-  std::chrono::steady_clock::time_point start_time;
-  std::chrono::steady_clock::time_point end_time;
-};
-
-//int main() {
-//  StopWatch sw;
-//  sw.start();
-//  // Simulate some work by sleeping for a specific duration
-//  std::this_thread::sleep_for(std::chrono::seconds(1));
-//  sw.stop();
-//  sw.print(std::cout); // This will print the elapsed time in the specified format
-//  return 0;
-//}
+namespace
+{
 
 }
 
@@ -86,13 +45,13 @@ Uuid ReadVtkStructuredPointsFilter::uuid() const
 //------------------------------------------------------------------------------
 std::string ReadVtkStructuredPointsFilter::humanName() const
 {
-  return "VTK STRUCTURED_POINTS Importer";
+  return "Read Vtk Structured Points File";
 }
 
 //------------------------------------------------------------------------------
 std::vector<std::string> ReadVtkStructuredPointsFilter::defaultTags() const
 {
-  return {className(), "IO", "Input", "Read", "Import"};
+  return {className(), "IO", "Input", "Read", "Import", "Vtk", "Structured Points"};
 }
 
 //------------------------------------------------------------------------------
@@ -102,8 +61,8 @@ Parameters ReadVtkStructuredPointsFilter::parameters() const
   // Create the parameter descriptors that are needed for this filter
   params.insertSeparator(Parameters::Separator{"Input Parameters"});
 
-  params.insert(std::make_unique<FileSystemPathParameter>(k_InputFile_Key, "Input VTK File", "The path to the input file", fs::path("DefaultInputFileName"),
-                                                          FileSystemPathParameter::ExtensionsType{".vtk"}, FileSystemPathParameter::PathType::InputFile));
+  params.insert(std::make_unique<FileSystemPathParameter>(k_InputFile_Key, "Input VTK File", "The path to the input file", fs::path("data.vtk"), FileSystemPathParameter::ExtensionsType{".vtk"},
+                                                          FileSystemPathParameter::PathType::InputFile));
   params.insertLinkableParameter(std::make_unique<BoolParameter>(k_ReadPointData_Key, "Read Point Data", "", false));
   params.insertLinkableParameter(std::make_unique<BoolParameter>(k_ReadCellData_Key, "Read Cell Data", "", false));
 
@@ -145,15 +104,24 @@ IFilter::PreflightResult ReadVtkStructuredPointsFilter::preflightImpl(const Data
   nx::core::Result<OutputActions> resultOutputActions;
   std::vector<PreflightValue> preflightUpdatedValues;
 
-  StopWatch sw;
+  nx::core::StopWatch sw;
   sw.start();
-  VtkLegacyFileReader legacyReader(pInputFileValue);
-  legacyReader.setPreflight(true);
-  resultOutputActions = legacyReader.preflightFile(pReadPointDataValue, pReadCellDataValue, pVertexGeometryPath, pImageGeometryPath, pVertexAttributeMatrixNameValue, pCellAttributeMatrixNameValue);
+
+  ReadVtkStructuredPointsInputValues inputValues;
+
+  inputValues.InputFile = filterArgs.value<FileSystemPathParameter::ValueType>(k_InputFile_Key);
+  inputValues.ReadPointData = filterArgs.value<bool>(k_ReadPointData_Key);
+  inputValues.ReadCellData = filterArgs.value<bool>(k_ReadCellData_Key);
+  inputValues.PointGeomPath = filterArgs.value<DataPath>(k_VertexDataContainerName_Key);
+  inputValues.ImageGeomPath = filterArgs.value<DataPath>(k_VolumeDataContainerName_Key);
+  inputValues.PointAttributeMatrixName = filterArgs.value<DataObjectNameParameter::ValueType>(k_VertexAttributeMatrixName_Key);
+  inputValues.CellAttributeMatrixName = filterArgs.value<DataObjectNameParameter::ValueType>(k_CellAttributeMatrixName_Key);
+
+  resultOutputActions = ReadVtkStructuredPoints::PreflightFile(inputValues);
+
   sw.stop();
   sw.print(std::cout);
   std::cout << "\n";
-
 
   // Return both the resultOutputActions and the preflightUpdatedValues via std::move()
   return {std::move(resultOutputActions), std::move(preflightUpdatedValues)};
@@ -168,8 +136,8 @@ Result<> ReadVtkStructuredPointsFilter::executeImpl(DataStructure& dataStructure
   inputValues.InputFile = filterArgs.value<FileSystemPathParameter::ValueType>(k_InputFile_Key);
   inputValues.ReadPointData = filterArgs.value<bool>(k_ReadPointData_Key);
   inputValues.ReadCellData = filterArgs.value<bool>(k_ReadCellData_Key);
-  inputValues.PointGeometryPath = filterArgs.value<DataPath>(k_VertexDataContainerName_Key);
-  inputValues.CellGeometryPath = filterArgs.value<DataPath>(k_VolumeDataContainerName_Key);
+  inputValues.PointGeomPath = filterArgs.value<DataPath>(k_VertexDataContainerName_Key);
+  inputValues.ImageGeomPath = filterArgs.value<DataPath>(k_VolumeDataContainerName_Key);
   inputValues.PointAttributeMatrixName = filterArgs.value<DataObjectNameParameter::ValueType>(k_VertexAttributeMatrixName_Key);
   inputValues.CellAttributeMatrixName = filterArgs.value<DataObjectNameParameter::ValueType>(k_CellAttributeMatrixName_Key);
